my_project/requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-bootstrap4==22.2
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
kombu==5.2.4
Pillow==9.3.0
postgres==4.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycparser==2.21
pymemcache==4.0.0
python-memcached==1.59
pytz==2022.6
rabbitmq==0.2.0
send-email==201212
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
